add integrations for taxonomy registrar
